api/requirements.txt
flask
celery
flask_sqlalchemy
redis
python-dotenv
psycopg2
flask-cors
bcrypt
black
openai